implement to_memconfig
this function converts a BlacksmithConfig to a MemConfiguration for use with DRAMAddr. For convenience, we'll use
std::array instead of C arrays in MemConfiguration. We also define MTX_SIZE as an unsigned constant to silence those
pesky compiler warnings. For matrix inversion (DRAM_MTX -> ADDR_MTX), we use the Eigen library.